Change Dag behavior significantly
Dags now never change the objects added to them; not even the inputs. Operation
inputs are now always bytes. This makes the dag central to finding connections.

Removed the concept of the MemoryDag; all dags are now set subclasses.

Related to these changes is that Dag.path() now will find a path even if start
isn't in the dag, but is an input of an operation in the dag.

Finally re-worked terminology to refer to operations rather than digests in
most places.