Make trailing garbage after attestations an error
Previously extra trailing data didn't raise a deserialization error and was
ignored, but it also wasn't round-trip serialized/deserialized making it not
very useful for protocol upgrades. For the pending and bitcoin attestations I
think there's a good case to be made that we're not going to do any future
upgrades by adding extra data, so better to just leave this out of the
protocol.