Display Bitcoin timestamps with appropriate precision
Previously would display them with precision down to the second, which
misrepresents how precise a Bitcoin timestamp can actually be, as
adversarial miners can get away with creating blocks whose timestamps
are inaccurate by multiple hours or more.

For those who do want a more precise timestamp, the height of the block
attesting to the timestamp is now displayed, allowing a manual
investigation of it.