Implement watchlist cache and runtime cache
A watchlist cache is implemented. The watchlist is a list of immature
timestamp files that should be upgraded at a later time. When a file is
timestamped, its file path is added to the watchlist. The file is
removed from the watchlis tonce it has been successfully pgraded.

Immature timestamps no longer immediately exits the program with error,
and instead an error is printed to stderr followed by a continuance of
the program until the very end when the program exits with exit(1).

A runtime cache is implemented. It is based on the premise that it is
unlikely that a timestamp will be upgraded in the time that it takes to
run the whole command, so the calendar servers will only be checked once
per each new commitment. This is particularly useful for bulk upgrading
of timestamps. If "immediate exit" was still implemented, this would not
be possible. Without a runtime cache, checking each timestamp against
each calendar server-- even with the same root commitment-- would be
unbearably slow.