🩹 Fix creation of invalid PGP signatures with git options
Options directly passed to git (like `git -c config.bla` or
`git --git-dir=...`) caused the first argument to show up before the PGP
signature, breaking git {log,show} --show-signature and thus also OTS
verification. This was especially annyoing as `git annex sync` or
`datalad save` pass options like this to git when committing.

With quite some `strace`ing and debugging I was finally
able to track down this bug to this missing `-q` to that one grep in the
OTS shell wrapper 🤦

It might be a good idea to ditch the shell wrapper and use a proper
Python for this. Wouldn't have happened with Python...